Move get IP address to app front end and show multiple IPs
